Create example to customize file adapter to read excel file

+++ bin/update_package.js
@@ -1,15 +1,20 @@
-const { writeFileSync, readFileSync, copyFileSync } = require('fs')
+const { writeFileSync, readFileSync, copyFileSync, existsSync } = require('fs')
 const { join } = require('path')
 
-new Array('package.json', 'yarn.lock', 'schema.json', '.npmignore', 'README.md', 'GUIDE.md')
-  .forEach(file => {
-    const fout = join(__dirname, '..', 'dist', file)
-    const fin = join(__dirname, '..', file)
-    if (file === 'package.json') {
-      let json = readFileSync(fin).toString()
-      json = json.replace(/^\s*"prepare"\:.+\n/m, '')
-      writeFileSync(fout, json)
-      return
-    }
-    copyFileSync(fin, fout)
-  })
+const flout = join(__dirname, '..', 'dist')
+if (existsSync(flout)) {
+  new Array('package.json', 'yarn.lock', 'schema.json', '.npmignore', 'README.md', 'GUIDE.md')
+    .forEach(file => {
+      const fin = join(__dirname, '..', file)
+      if (existsSync(fin)) {
+        const fout = join(flout, file)
+        if (file === 'package.json') {
+          let json = readFileSync(fin).toString()
+          json = json.replace(/^\s*"prepare"\:.+\n/m, '')
+          writeFileSync(fout, json)
+          return
+        }
+        copyFileSync(fin, fout)
+      }
+    })
+}
